Moves the auth module into the publish package
